Move py3 tests to cover
and simplify scripts accordingly
