+++ hypothesis-python/scripts/basic-test.sh
@@ -16,7 +16,7 @@ pip install .
 PYTEST="python -m pytest -n2"
 
 # Run all the no-extra-dependency tests for this version (except slow nocover tests)
-$PYTEST tests/cover tests/pytest tests/py3
+$PYTEST tests/cover tests/pytest
 
 # Run tests for each extra module while the requirements are installed
 pip install ".[pytz, dateutil]"
