+++ hypothesis-python/tests/README.rst
@@ -102,11 +102,6 @@ Still here?  Here's a note on what to expect in each directory.
 ``pandas/``
     Tests for the Pandas extra.
 
-``py3/``
-    Tests that require Python 3.  Includes checking that unicode identifiers
-    and function annotations don't break anything, asyncio tests, and tests
-    for inference from type hints.
-
 ``pytest/``
     Hypothesis has excellent integration with ``pytest``, though we are careful
     to support other test runners including unittest and nose.  This is where we
