--- hypothesis-python/tests/py3/test_falsifying_example_output.py
@@ -1,35 +1,0 @@
-# coding=utf-8
-#
-# This file is part of Hypothesis, which may be found at
-# https://github.com/HypothesisWorks/hypothesis/
-#
-# Most of this work is copyright (C) 2013-2019 David R. MacIver
-# (david@drmaciver.com), but it contains contributions by others. See
-# CONTRIBUTING.rst for a full list of people who may hold copyright, and
-# consult the git log if you need to determine who owns an individual
-# contribution.
-#
-# This Source Code Form is subject to the terms of the Mozilla Public License,
-# v. 2.0. If a copy of the MPL was not distributed with this file, You can
-# obtain one at https://mozilla.org/MPL/2.0/.
-#
-# END HEADER
-
-from __future__ import absolute_import, division, print_function
-
-import pytest
-
-from hypothesis import given, strategies as st
-from tests.common.utils import capture_out
-
-
-def test_vararg_output():
-    @given(foo=st.text())
-    def test(*args, foo):
-        assert False
-
-    with capture_out() as cap:
-        with pytest.raises(AssertionError):
-            test(1, 2, 3)
-
-    assert "1, 2, 3" in cap.getvalue()
