+++ hypothesis-python/tox.ini
@@ -121,7 +121,7 @@ commands =
     rm -f branch-check
     python -m coverage --version
     python -m coverage debug sys
-    python -m coverage run --rcfile=.coveragerc -m pytest -n0 --strict tests/cover tests/conjecture tests/datetime tests/py3 tests/numpy tests/pandas tests/lark --ff {posargs}
+    python -m coverage run --rcfile=.coveragerc -m pytest -n0 --strict tests/cover tests/conjecture tests/datetime tests/numpy tests/pandas tests/lark --ff {posargs}
     python -m coverage report -m --fail-under=100 --show-missing --skip-covered
     python scripts/validate_branch_check.py
